refactor: 로또 생성 전략 패키지명 및 클래스명 변경
