+++ src/test/java/domain/LottoStoreTest.java
@@ -2,7 +2,7 @@ package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.numberstrategy.NumberPickStrategy;
+import domain.lottogeneratestrategy.LottoPickStrategy;
 import java.util.List;
 import java.util.Set;
 import org.junit.jupiter.api.Test;
@@ -13,7 +13,7 @@ class LottoStoreTest {
     void 돈을_이용해_로또를_살_수_있다() {
         //given
         Money money = new Money(3000);
-        NumberPickStrategy fixNumberStrategy = (int maxNumber, int size) -> List.of(1, 2, 3, 4, 5, 6);
+        LottoPickStrategy fixNumberStrategy = (int maxNumber, int size) -> List.of(1, 2, 3, 4, 5, 6);
         LottoMachine lottoMachine = new LottoMachine(fixNumberStrategy);
         LottoStore lottoStore = new LottoStore(lottoMachine);
